fixed crap code
rewrite in ts
decent formatting
use consistant functions
correct spacing
make the code look ok
