+++ tsconfig.json
@@ -1,0 +1,19 @@
+{
+	"compilerOptions": {
+		"lib": ["ES2018"],
+		"module": "commonjs",
+		"moduleResolution": "node",
+		"target": "ES2018",
+		"outDir": "dist",
+		"sourceMap": true,
+		"esModuleInterop": true,
+		"experimentalDecorators": true,
+		"emitDecoratorMetadata": true,
+		"allowSyntheticDefaultImports": true,
+		"skipLibCheck": true,
+		"skipDefaultLibCheck": true,
+		"declaration": true
+	},
+	"include": ["src"],
+	"exclude": ["node_modules", "**/*.spec.ts"]
+}
